pyproject.toml
[project]
name = "tripsync-backend"
version = "0.1.0"
description = "backend for tripsync"
authors = [{name = "Pushkar Singhal",email = "pusinghal010@gmail.com"}]
readme = "README.md"

requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "uvicorn[standard] (>=0.43.0,<0.44.0)",
    "adrf (>=0.1.12,<0.2.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "redis[asyncio] (>=7.4.0,<8.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "celery[rabbitmq] (>=5.6.3,<6.0.0)",
    "whitenoise (>=6.12.0,<7.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0"
pytest-django = ">=4.8.0"                              
fakeredis = {extras = ["aioredis"], version = "^2.34.1"}
 
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "main.settings"
python_files = ["tests.py"]
python_classes = ["*Tests"]
python_functions = ["test_*"]